Preload the fetch data
This works OK and helps speed things up (the time to download JS,
excecute, fetch is ~equal to the time to fetch, so we parallelise
it). This has the downside of baking in the location we are fetching
to the initial HTML. This won't work beyond only Helsinki, and also
is not great if we ever cache the responses in a service worker.

Anyway, this should help for now.